drm/i915/gvt: Unconditionally put reference to KVM when detaching vGPU
Always put the KVM reference when closing a vCPU device, as
intel_vgpu_open_device() succeeds if and only if the KVM pointer is
valid and a reference to KVM is acquired.  And if that doesn't hold true,
the call to kvm_page_track_unregister_notifier() a few lines earlier is
doomed.

Signed-off-by: Sean Christopherson <seanjc@google.com>
Reviewed-by: Kevin Tian <kevin.tian@intel.com>