KVM: x86: SVM: Add VMX style functions for SVM MSR management
VMX MSR interception is done via three functions:

        vmx_disable_intercept_for_msr(vcpu, msr, type)
        vmx_enable_intercept_for_msr(vcpu, msr, type)
        vmx_set_intercept_for_msr(vcpu, msr, type, value)

While SVM uses

        set_msr_interception(vcpu, msrpm, msr, read, write)

The SVM code is unintuitive (using 0 for enable and 1 for disable), and
forces both read and write changes with each call which is not always
required.

Add helpers functions to SVM to match VMX:

        svm_disable_intercept_for_msr(vcpu, msr, type)
        svm_enable_intercept_for_msr(vcpu, msr, type)
        svm_set_intercept_for_msr(vcpu, msr, type, enable_intercept)

Update calls to set_msr_interception() to use the new functions. This
update is only made to calls that toggle interception for both read and
write.

Keep the old paths for now, they will be deleted once all code is
converted to the new helpers.

No functional change.

Suggested-by: Sean Christopherson <seanjc@google.com>
Signed-off-by: Anish Ghulati <aghulati@google.com>