KVM: selftests: Add kvm_snprintf() to KVM selftests
Add a local version of kvm_snprintf() for use in the guest.

Having a local copy allows the guest access to string formatting options
without dependencies on LIBC.  LIBC is problematic because it heavily
relies on both AVX-512 instructions and a TLS, neither of which are
guaranteed to be set up in the guest.  Using tools' nolibc would work in
theory, but would be significantly more work and may or may not be an
overall improvement.

The file kvm_sprintf.c was lifted from arch/x86/boot/printf.c and
adapted to work in the guest, including the addition of buffer length.
I.e. s/sprintf/snprintf/

Prefix the new functions with "kvm_" to allow selftests (guests in
particular) to explicitly call them, i.e. to ensure the guest-friendly
version is used.

Link: https://lore.kernel.org/all/mtdi6smhur5rqffvpu7qux7mptonw223y2653x2nwzvgm72nlo@zyc4w3kwl3rg
Signed-off-by: Aaron Lewis <aaronlewis@google.com>
Signed-off-by: Sean Christopherson <seanjc@google.com>