vfio: Assert that device is fully closed when putting KVM reference
WARN if a device is not fully closed, i.e. if close_device() hasn't yet
been called, when (potentially) putting the reference to KVM that was
acquried on behalf of the device.  All downstream device implementations
that utilize KVM rely on the core VFIO device framework to handle pinning
the KVM *module*.

Specifically, s390's adjunct processor matrix and x86's KVMGT both grab
references to "struct kvm", but neither grabs an additional reference to
kvm_put_kvm(), i.e. don't ensure that the KVM module is pinned, and so
subtly rely on core VFIO to keep KVM alive.

Signed-off-by: Sean Christopherson <seanjc@google.com>