*** DO NOT APPLY *** KVM: x86: Verify KVM initializes all consumed guest caps
Assert that all features accessed outside of kvm_set_cpu_caps() are known
to KVM, i.e. that KVM doesn't set/clear a feature that wasn't documented,
and that KVM doesn't query a feature that can never actually be set.

Do not apply, as the proper way to enforce this is to do post-processing at
build time (and there are other shortcomings of this PoC, e.g. it requires
all KVM modules to be built-in).

Not-signed-off-by: Sean Christopherson <seanjc@google.com>