KVM: SVM: Leave IsRunning=1 when the vCPU is put, but not blocking
When unloading a vCPU, keep IsRunning set in both the table used for
virtual IPIs, and in IRTEs associated with the vCPU.  KVM relies on the
VM-Exit (IPIs) and GA log interrupt (devices) only for blocking vCPUs,
i.e. if and only KVM needs to be made aware that the vCPU has a pending
wake event.  Leaving IsRunning=1 means other vCPUs don't need to VM-Exit
when sending an IPI to vCPUs that have exited to userspace or been
scheduled out, and similarly

Signed-off-by: Sean Christopherson <seanjc@google.com>